refactor: product types push
